refactor: moved usePixiApp to hooks module; separated shared values/functions to new stageProps.js module
