+++ package.json
@@ -7,15 +7,7 @@
   "license": "MIT",
   "repository": "github:michalochman/react-pixi-fiber",
   "typings": "./index.d.ts",
-  "files": [
-    "LICENSE",
-    "README.md",
-    "index.d.ts",
-    "index.js",
-    "react-pixi-alias.js",
-    "cjs/",
-    "src/"
-  ],
+  "files": ["LICENSE", "README.md", "index.d.ts", "index.js", "react-pixi-alias.js", "cjs/", "src/"],
   "dependencies": {
     "fbjs": "^0.8.0",
     "performance-now": "^2.1.0",
@@ -50,9 +42,9 @@
     "jest": "^22.4.3",
     "pixi.js": "^4.4.0",
     "prettier": "^1.12.1",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
-    "react-test-renderer": "^16.0.0",
+    "react": "^16.7.0-alpha.2",
+    "react-dom": "^16.7.0-alpha.2",
+    "react-test-renderer": "^16.7.0-alpha.2",
     "rollup": "^0.58.2",
     "rollup-plugin-babel": "^3.0.4",
     "rollup-plugin-commonjs": "^9.1.3",
@@ -80,19 +72,12 @@
   },
   "jest": {
     "coverageDirectory": "coverage",
-    "coverageReporters": [
-      "lcov",
-      "text-summary"
-    ],
-    "collectCoverageFrom": [
-      "src/**/*.js"
-    ],
+    "coverageReporters": ["lcov", "text-summary"],
+    "collectCoverageFrom": ["src/**/*.js"],
     "globals": {
       "__DEV__": true
     },
-    "setupFiles": [
-      "./test/jest.setupPixi.js"
-    ],
+    "setupFiles": ["./test/jest.setupPixi.js"],
     "transform": {
       "^.+\\.js$": "babel-jest"
     },
