+++ src/hooks.js
@@ -1,5 +1,8 @@
+import React from "react";
+import { useContext, useEffect, useRef, useState, useLayoutEffect } from "react";
 import { AppContext } from "./AppProvider";
-import { useContext, useEffect, useRef } from "react";
+import { getCanvasProps } from "./stageProps";
+import { unmount } from "./render";
 
 export function usePixi() {
   return useContext(AppContext);
@@ -25,4 +28,33 @@ export function usePreviousProps(value) {
   });
 
   return ref.current;
+}
+
+export function usePixiApp(props) {
+  const { options, width, height } = props;
+  const canvasRef = useRef(null);
+  const [app, setApp] = useState(null);
+  const canvasProps = getCanvasProps(props);
+  // Do not render anything if view is passed to options
+  const canvas = options && options.view ? null : <canvas ref={canvasRef} {...canvasProps} />;
+
+  // Initialize pixi application on mount
+  useLayoutEffect(() => {
+    const appInstance = new PIXI.Application({
+      height,
+      width,
+      view: canvasRef.current,
+      ...options,
+    });
+
+    setApp(appInstance);
+
+    // Destroy pixi application on unmount
+    return () => {
+      unmount(appInstance.stage);
+      appInstance.destroy();
+    };
+  }, []);
+
+  return { app, canvas };
 }
