+++ test/StageLegacy.test.js
@@ -1,9 +1,15 @@
+jest.mock("react", () =>
+  Object.assign(require.requireActual("react"), {
+    useState: null,
+  })
+);
+
 import React from "react";
 import renderer from "react-test-renderer";
 import * as PIXI from "pixi.js";
 import { Text } from "../src/index";
 import ReactPixiFiber from "../src/ReactPixiFiber";
-import Stage, { appTestHook } from "../src/Stage";
+import Stage from "../src/Stage";
 import {
   getCanvasProps,
   getDisplayObjectProps,
@@ -30,7 +36,7 @@ jest.mock("../src/render", () => {
   };
 });
 
-describe("Stage (function)", () => {
+describe("Stage (class)", () => {
   it("renders canvas element", () => {
     const tree = renderer.create(<Stage />).toJSON();
 
@@ -78,9 +84,10 @@ describe("Stage (function)", () => {
 
     const element = renderer.create(<Stage height={height} width={width} options={options} />);
     const instance = element.getInstance();
+    const app = instance._app;
 
-    expect(appTestHook instanceof PIXI.Application).toBeTruthy();
-    expect(appTestHook._options).toMatchObject({
+    expect(app instanceof PIXI.Application).toBeTruthy();
+    expect(app._options).toMatchObject({
       height,
       width,
       ...options,
@@ -97,15 +104,17 @@ describe("Stage (function)", () => {
 
     const element = renderer.create(<Stage options={options} />);
     const instance = element.getInstance();
+    const app = instance._app;
 
-    expect(appTestHook instanceof PIXI.Application).toBeTruthy();
-    expect(appTestHook._options).toMatchObject(options);
+    expect(app instanceof PIXI.Application).toBeTruthy();
+    expect(app._options).toMatchObject(options);
   });
 
   it("creates root Container", () => {
     const element = renderer.create(<Stage height={300} width={400} scale={2} position="40,20" />);
     const instance = element.getInstance();
-    const stage = appTestHook.stage;
+    const app = instance._app;
+    const stage = app.stage;
 
     expect(stage instanceof PIXI.Container).toBeTruthy();
   });
@@ -116,7 +125,8 @@ describe("Stage (function)", () => {
     const y = 20;
     const element = renderer.create(<Stage height={300} position={`${x},${y}`} scale={scale} width={400} />);
     const instance = element.getInstance();
-    const stage = appTestHook.stage;
+    const app = instance._app;
+    const stage = app.stage;
 
     expect(stage.position.x).toEqual(x);
     expect(stage.position.y).toEqual(y);
@@ -128,7 +138,8 @@ describe("Stage (function)", () => {
     const scale = 2;
     const element = renderer.create(<Stage height={300} scale={scale} width={400} />);
     const instance = element.getInstance();
-    const stage = appTestHook.stage;
+    const app = instance._app;
+    const stage = app.stage;
 
     expect(stage.scale.x).toEqual(scale);
     expect(stage.scale.y).toEqual(scale);
@@ -145,16 +156,17 @@ describe("Stage (function)", () => {
     const width = 400;
     const element = renderer.create(<Stage height={height} width={width} />);
     const instance = element.getInstance();
+    const app = instance._app;
 
-    expect(appTestHook.renderer.height).toEqual(height);
-    expect(appTestHook.renderer.width).toEqual(width);
+    expect(app.renderer.height).toEqual(height);
+    expect(app.renderer.width).toEqual(width);
 
     const newHeight = 600;
     const newWidth = 800;
     element.update(<Stage height={newHeight} width={newWidth} />);
 
-    expect(appTestHook.renderer.height).toEqual(newHeight);
-    expect(appTestHook.renderer.width).toEqual(newWidth);
+    expect(app.renderer.height).toEqual(newHeight);
+    expect(app.renderer.width).toEqual(newWidth);
   });
 
   it("can be umounted", () => {
@@ -168,31 +180,30 @@ describe("Stage (function)", () => {
     const children = <Text text="Hello World!" />;
     const element = renderer.create(<Stage>{children}</Stage>);
     const instance = element.getInstance();
-    const stage = appTestHook.stage;
+    const stage = instance._app.stage;
 
     expect(render).toHaveBeenCalledTimes(1);
-    expect(render).toHaveBeenCalledWith(<AppProvider app={appTestHook}>{children}</AppProvider>, stage);
+    expect(render).toHaveBeenCalledWith(<AppProvider app={instance._app}>{children}</AppProvider>, stage);
   });
 
   it("calls render on componentDidUpdate", () => {
     const children1 = <Text text="Hello World!" />;
     const element = renderer.create(<Stage>{children1}</Stage>);
     const instance = element.getInstance();
-    const stage = appTestHook.stage;
+    const stage = instance._app.stage;
 
     render.mockClear();
     const children2 = <Text text="World Hello!" />;
     element.update(<Stage>{children2}</Stage>);
 
     expect(render).toHaveBeenCalledTimes(1);
-    expect(render).toHaveBeenCalledWith(<AppProvider app={appTestHook}>{children2}</AppProvider>, stage);
+    expect(render).toHaveBeenCalledWith(<AppProvider app={instance._app}>{children2}</AppProvider>, stage);
   });
 
   it("calls unmount on componentWillUnmount", () => {
     const element = renderer.create(<Stage />);
     const instance = element.getInstance();
-    const { stage } = appTestHook;
-
+    const stage = instance._app.stage;
     unmount.mockClear();
     element.unmount();
